requirements.txt
aiohttp==3.7.4.post0
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.4
backcall==0.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
colorama==0.4.4
decorator==4.4.2
discord.py==1.7.3
et-xmlfile==1.1.0
idna==3.3
iniconfig==1.1.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.2.0
numpy==1.21.4
openai==0.11.0
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.39
parso==0.8.1
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.10
py==1.10.0
pycodestyle==2.6.0
Pygments==2.7.4
pylint==2.6.0
pytest==6.2.1
python-dateutil==2.8.1
pytz==2021.3
pywin32==300
pyzmq==21.0.0
requests==2.26.0
rope==0.18.0
six==1.15.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
wincertstore==0.2
wrapt==1.12.1
yarl==1.7.2
